[refactor] Move call center options view to call center app
